+++ lib/truffle/truffle/cext.rb
@@ -412,8 +412,12 @@ module Truffle::CExt
     ruby_class
   end
 
-  def rb_obj_respond_to(object, id, priv)
-    Primitive.object_respond_to?(object, id, priv != 0)
+  def rb_obj_respond_to(object, name, priv)
+    object.respond_to?(name, priv != 0)
+  end
+
+  def rb_respond_to(object, name)
+    object.respond_to?(name, false)
   end
 
   def rb_check_convert_type(obj, type_name, method)
@@ -840,10 +844,6 @@ module Truffle::CExt
 
   def rb_funcall_with_block(recv, meth, args, block)
     recv.public_send(meth, *args, &block)
-  end
-
-  def rb_respond_to(object, name)
-    object.respond_to?(name)
   end
 
   def rb_funcallv_public(recv, meth, args)
@@ -1700,8 +1700,7 @@ module Truffle::CExt
   end
 
   def rb_backref_get
-    Primitive.frame_local_variable_get(:$~,
-      Truffle::ThreadOperations.ruby_caller([Truffle::CExt, Truffle::Interop.singleton_class]))
+    Primitive.regexp_last_match_get(Truffle::ThreadOperations.ruby_caller_special_variables([Truffle::CExt, Truffle::Interop.singleton_class]))
   end
 
   def rb_gv_set(name, value)
@@ -1719,7 +1718,7 @@ module Truffle::CExt
 
   def rb_reg_match(re, str)
     result = str ? Truffle::RegexpOperations.match(re, str, 0) : nil
-    Primitive.frame_local_variable_set(:$~, result, Truffle::ThreadOperations.ruby_caller([Truffle::CExt, Truffle::Interop.singleton_class]))
+    Primitive.regexp_last_match_set(Truffle::ThreadOperations.ruby_caller_special_variables([Truffle::CExt, Truffle::Interop.singleton_class]), result)
 
     result.begin(0) if result
   end