+++ src/main/java/org/truffleruby/core/CoreLibrary.java
@@ -25,6 +25,7 @@ import java.util.concurrent.ConcurrentMap;
 
 import org.jcodings.specific.USASCIIEncoding;
 import org.jcodings.transcode.EConvFlags;
+import org.truffleruby.Layouts;
 import org.truffleruby.RubyContext;
 import org.truffleruby.RubyLanguage;
 import org.truffleruby.SuppressFBWarnings;
@@ -46,6 +47,8 @@ import org.truffleruby.core.rope.Rope;
 import org.truffleruby.core.string.RubyString;
 import org.truffleruby.core.string.StringOperations;
 import org.truffleruby.core.symbol.RubySymbol;
+import org.truffleruby.debug.GlobalVariablesObject;
+import org.truffleruby.debug.TopScopeObject;
 import org.truffleruby.extra.ffi.Pointer;
 import org.truffleruby.language.Nil;
 import org.truffleruby.language.NotProvided;
@@ -77,6 +80,7 @@ import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.TruffleOptions;
 import com.oracle.truffle.api.frame.FrameDescriptor;
+import com.oracle.truffle.api.frame.FrameSlot;
 import com.oracle.truffle.api.object.DynamicObjectLibrary;
 import com.oracle.truffle.api.source.Source;
 import com.oracle.truffle.api.source.SourceSection;
@@ -224,6 +228,11 @@ public class CoreLibrary {
     public final GlobalVariables globalVariables;
 
     public final FrameDescriptor emptyDescriptor;
+    /* Some things (such as procs created from symbols) require a declaration frame, and this should include a slot for
+     * special variable storage. This frame descriptor should be used for those frames to provide a constant frame
+     * descriptor in those cases. */
+    public final FrameDescriptor emptyDeclarationDescriptor;
+    public final FrameSlot emptyDeclarationSpecialVariableSlot;
 
     @CompilationFinal private RubyClass eagainWaitReadable;
     @CompilationFinal private RubyClass eagainWaitWritable;
@@ -242,6 +251,7 @@ public class CoreLibrary {
     @CompilationFinal private GlobalVariableReader stderrReader;
 
     @CompilationFinal public RubyBinding topLevelBinding;
+    @CompilationFinal public TopScopeObject topScopeObject;
 
     private final ConcurrentMap<String, Boolean> patchFiles;
 
@@ -585,9 +595,14 @@ public class CoreLibrary {
 
         mainObject = new RubyBasicObject(objectClass, language.basicObjectShape);
         emptyDescriptor = new FrameDescriptor(Nil.INSTANCE);
+        emptyDeclarationDescriptor = new FrameDescriptor(Nil.INSTANCE);
+        emptyDeclarationSpecialVariableSlot = emptyDeclarationDescriptor
+                .addFrameSlot(Layouts.SPECIAL_VARIABLES_STORAGE);
         argv = new RubyArray(arrayClass, RubyLanguage.arrayShape, ArrayStoreLibrary.INITIAL_STORE, 0);
 
         globalVariables = new GlobalVariables();
+        topScopeObject = new TopScopeObject(
+                new Object[]{ new GlobalVariablesObject(globalVariables), mainObject });
 
         patchFiles = initializePatching(context);
     }