+++ src/main/java/org/truffleruby/core/proc/ProcNodes.java
@@ -72,12 +72,12 @@ public abstract class ProcNodes {
 
         @Specialization
         protected RubyProc proc(VirtualFrame frame, RubyClass procClass, Object[] args, NotProvided block,
-                @Cached("create(nil)") FindAndReadDeclarationVariableNode readNode,
+                @Cached FindAndReadDeclarationVariableNode readNode,
                 @Cached ReadCallerFrameNode readCaller,
                 @Cached ProcNewNode recurseNode) {
             final MaterializedFrame parentFrame = readCaller.execute(frame);
 
-            Object parentBlock = readNode.execute(parentFrame, TranslatorEnvironment.METHOD_BLOCK_NAME);
+            Object parentBlock = readNode.execute(parentFrame, TranslatorEnvironment.METHOD_BLOCK_NAME, nil);
 
             if (parentBlock == nil) {
                 throw new RaiseException(getContext(), coreExceptions().argumentErrorProcWithoutBlock(this));
@@ -112,6 +112,7 @@ public abstract class ProcNodes {
                     block.callTargetForType,
                     block.callTargetForLambdas,
                     block.declarationFrame,
+                    block.declarationStorage,
                     block.method,
                     block.block,
                     block.frameOnStackMarker,
@@ -151,6 +152,7 @@ public abstract class ProcNodes {
                     proc.callTargetForType,
                     proc.callTargetForLambdas,
                     proc.declarationFrame,
+                    proc.declarationStorage,
                     proc.method,
                     proc.block,
                     proc.frameOnStackMarker,
@@ -272,6 +274,7 @@ public abstract class ProcNodes {
                     block.callTargetForType,
                     block.callTargetForLambdas,
                     block.declarationFrame,
+                    block.declarationStorage,
                     block.method,
                     block.block,
                     block.frameOnStackMarker,