+++ src/main/java/org/truffleruby/core/support/TypeNodes.java
@@ -55,7 +55,6 @@ import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.ImportStatic;
 import com.oracle.truffle.api.dsl.NodeChild;
 import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.library.CachedLibrary;
 import com.oracle.truffle.api.object.DynamicObjectLibrary;
 import com.oracle.truffle.api.object.HiddenKey;
@@ -83,8 +82,9 @@ public abstract class TypeNodes {
                 .create(null, null, null);
 
         @Specialization
-        protected boolean objectRespondTo(VirtualFrame frame, Object object, Object name, boolean includePrivate) {
-            return respondToNode.executeDoesRespondTo(frame, object, name, includePrivate);
+        protected boolean objectRespondTo(Object object, Object name, boolean includePrivate) {
+            // Do not pass a frame here, we want to ignore refinements and not need to read the caller frame
+            return respondToNode.executeDoesRespondTo(null, object, name, includePrivate);
         }
 
     }
@@ -260,6 +260,12 @@ public abstract class TypeNodes {
     @Primitive(name = "object_can_contain_object")
     @ImportStatic(ArrayGuards.class)
     public abstract static class CanContainObjectNode extends PrimitiveArrayArgumentsNode {
+
+        public static CanContainObjectNode create() {
+            return TypeNodesFactory.CanContainObjectNodeFactory.create(null);
+        }
+
+        abstract public boolean execute(RubyArray array);
 
         @Specialization(
                 guards = {