+++ src/main/java/org/truffleruby/parser/MethodTranslator.java
@@ -12,10 +12,10 @@ package org.truffleruby.parser;
 import java.util.Arrays;
 
 import org.truffleruby.RubyContext;
+import org.truffleruby.collections.CachedSupplier;
 import org.truffleruby.core.IsNilNode;
 import org.truffleruby.core.cast.ArrayCastNodeGen;
 import org.truffleruby.core.proc.ProcType;
-import org.truffleruby.language.LazyRubyNode;
 import org.truffleruby.language.LexicalScope;
 import org.truffleruby.language.RubyNode;
 import org.truffleruby.language.RubyRootNode;
@@ -43,6 +43,7 @@ import org.truffleruby.language.supercall.ReadSuperArgumentsNode;
 import org.truffleruby.language.supercall.ReadZSuperArgumentsNode;
 import org.truffleruby.language.supercall.SuperCallNode;
 import org.truffleruby.language.supercall.ZSuperOutsideMethodNode;
+import org.truffleruby.language.threadlocal.MakeSpecialVariableStorageNode;
 import org.truffleruby.parser.ast.ArgsParseNode;
 import org.truffleruby.parser.ast.MethodDefParseNode;
 import org.truffleruby.parser.ast.ParseNode;
@@ -274,7 +275,10 @@ public class MethodTranslator extends BodyTranslator {
                 language,
                 arity,
                 sequence(bodySourceSection, Arrays.asList(loadArguments, body)));
+
         body.unsafeSetSourceSection(sourceSection);
+
+        body = sequence(bodySourceSection, Arrays.asList(new MakeSpecialVariableStorageNode(), body));
 
         if (environment.getFlipFlopStates().size() > 0) {
             body = sequence(bodySourceSection, Arrays.asList(initFlipFlopStates(sourceSection), body));
@@ -289,33 +293,32 @@ public class MethodTranslator extends BodyTranslator {
         return body;
     }
 
-    public RootCallTarget compileMethodNode(SourceIndexLength sourceSection, MethodDefParseNode defNode,
+    private RubyRootNode translateMethodNode(SourceIndexLength sourceSection, MethodDefParseNode defNode,
             ParseNode bodyNode) {
         final SourceIndexLength sourceIndexLength = defNode.getPosition();
         final SourceSection fullMethodSourceSection = sourceIndexLength.toSourceSection(source);
-
-        final RubyNode body;
-
-        if (shouldLazyTranslate) {
-            final TranslatorState state = getCurrentState();
-
-            body = new LazyRubyNode(language, () -> {
-                restoreState(state);
-                return compileMethodBody(sourceSection, bodyNode);
-            });
-        } else {
-            body = compileMethodBody(sourceSection, bodyNode);
-        }
-
-        final RubyRootNode rootNode = new RubyRootNode(
+        return new RubyRootNode(
                 context,
                 fullMethodSourceSection,
                 environment.getFrameDescriptor(),
                 environment.getSharedMethodInfo(),
-                body,
+                compileMethodBody(sourceSection, bodyNode),
                 Split.HEURISTIC);
+    }
 
-        return Truffle.getRuntime().createCallTarget(rootNode);
+    public CachedSupplier<RootCallTarget> buildMethodNodeCompiler(SourceIndexLength sourceSection,
+            MethodDefParseNode defNode, ParseNode bodyNode) {
+
+        if (shouldLazyTranslate) {
+            final TranslatorState state = getCurrentState();
+            return new CachedSupplier<>(() -> {
+                restoreState(state);
+                return Truffle.getRuntime().createCallTarget(translateMethodNode(sourceSection, defNode, bodyNode));
+            });
+        } else {
+            final RubyRootNode root = translateMethodNode(sourceSection, defNode, bodyNode);
+            return new CachedSupplier<>(() -> Truffle.getRuntime().createCallTarget(root));
+        }
     }
 
     private void declareArguments() {